+++ SPMSources/CASBaseResources/CleverAdsSolutionsResources.swift
@@ -1,0 +1,4 @@
+//
+//  CleverAdsSolutionsResources.swift
+//  CleverAdsSolutions
+//