+++ Adapters/AppLovin/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## AppLovin iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationAppLovin', '13.2.0.0'
+pod 'CASMediationAppLovin', '13.3.0.0'
 ```
+
+### 13.3.0.0
+- Certified with AppLovin - 13.3.0
+- Added Swift Package Manager integration support.
 
 ### 13.2.0.0
 - Certified with AppLovin - 13.2.0