+++ Adapters/DTExchange/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## DT Exchange iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationDTExchange', '8.3.6.0'
+pod 'CASMediationDTExchange', '8.3.7.0'
 ```
+
+### 8.3.7.0
+- Certified with DTExchange - 8.3.7
+- Added Swift Package Manager integration support.
 
 ### 8.3.6.0
 - Certified with DTExchange - 8.3.6