+++ Adapters/Kidoz/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## Kidoz iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationKidoz', '9.2.0.0'
+pod 'CASMediationKidoz', '10.0.2.0'
 ```
+
+### 10.0.2.0
+- Certified with Kidoz - 10.0.2
+- Added Swift Package Manager integration support.
 
 ### 9.2.0.0
 - Certified with Kidoz - 9.2.0