+++ Adapters/Prado/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## Prado Ads iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationPrado', '10.0.1.0'
+pod 'CASMediationPrado', '10.0.2.0'
 ```
+
+### 10.0.2.0
+- Certified with Prado - 10.0.2
+- Added Swift Package Manager integration support.
 
 ### 10.0.1.0
 - Certified with Prado - 10.0.1