+++ Adapters/UnityAds/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## Unity Ads iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationUnityAds', '4.14.2.0'
+pod 'CASMediationUnityAds', '4.15.0.0'
 ```
+
+### 4.15.0.0
+- Certified with UnityAds - 4.15.0
+- Added Swift Package Manager integration support.
 
 ### 4.14.2.0
 - Certified with UnityAds - 4.14.2