+++ Adapters/YandexAds/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## Yandex Ads iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationYandexAds', '7.12.1.0'
+pod 'CASMediationYandexAds', '7.13.0.0'
 ```
+
+### 7.13.0.0
+- Certified with YandexAds - 7.13.0
+- Added Swift Package Manager integration support.
 
 ### 7.12.1.0
 - Certified with YandexAds - 7.12.1