+++ Adapters/YsoNetwork/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## Yso Network iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationYsoNetwork', '1.1.31.0'
+pod 'CASMediationYsoNetwork', '1.1.31.1'
 ```
+
+### 1.1.31.1
+- Added Swift Package Manager integration support.
+- Fixed an issue with filling ads.
 
 ### 1.1.31.0
 - Certified with YsoNetwork - 1.1.31